Add FeaturedBox component Refactor FeaturedSection
