+++ src/components/common/FeaturedBox/FeaturedBox.module.scss
@@ -71,7 +71,7 @@
       align-items: flex-end;
       padding: 1rem 5px;
       height: 3.7rem;
-  
+
       div {
         background: $color-black;
         border-radius: 50%;
@@ -80,7 +80,7 @@
         color: $color-white;
         padding: 0.25rem 0;
         line-height: 1.6rem;
-  
+
         span {
           font-size: smaller;
         }
@@ -103,7 +103,7 @@ img {
 }
 .photo:hover .buttons__hover {
   visibility: visible;
-}  
+}
 
 
 .content {
