+++ src/components/common/FeaturedCarousel/FeaturedCarousel.js
