+++ src/components/common/FeaturedCarousel/FeaturedCarousel.module.scss
